+++ dss/api/bundles/__init__.py
@@ -13,7 +13,7 @@ from flask import jsonify, request
 
 from dss.storage.identifiers import TombstoneID, BundleFQID, FileFQID
 from dss.util.blobstore import test_object_exists, ObjectTest
-from dss.util.bundles import get_bundle
+from dss.storage.bundles import get_bundle
 from dss.util.version import datetime_to_version_format
 from dss import DSSException, dss_handler
 from dss.config import Config, Replica
